imageio: split HDRDecoder into its own file
