+++ libs/imageio/CMakeLists.txt
@@ -9,6 +9,7 @@ set(PUBLIC_HDR_DIR include)
 # ==================================================================================================
 set(PUBLIC_HDRS
         include/imageio/BlockCompression.h
+        include/imageio/HDRDecoder.h
         include/imageio/ImageDecoder.h
         include/imageio/ImageDiffer.h
         include/imageio/ImageEncoder.h
@@ -16,6 +17,7 @@ set(PUBLIC_HDRS
 
 set(SRCS
         src/BlockCompression.cpp
+        src/HDRDecoder.cpp
         src/ImageDecoder.cpp
         src/ImageDiffer.cpp
         src/ImageEncoder.cpp
